4.14

+++ part4/bloglist/controllers/blog.js
@@ -20,12 +20,27 @@ blogRouter.post('/', (request, response, next) => {
     })
     .catch(err => next(err))
 })
-blogRouter.delete('/:id', (req, res, next) => {
-  Blog.findByIdAndRemove(req.params.id)
-    .then((response) => {
+blogRouter.delete('/:id', async (req, res, next) => {
+  const blog = await Blog.findById(req.params.id)
+  if(!blog){
+    return res.status(404).json({error : "No data found"})
+  }
+
+   await Blog.findByIdAndRemove(req.params.id)
+   .then((response) => {
       console.log(response)
       return res.status(200).json(response)
     })
     .catch((error) => next(error))
+})
+
+blogRouter.put('/:id', async (req,res,next)=>{
+  const blog = await Blog.findById(req.params.id)
+  if(!blog){
+    return res.status(404).json({error : "No data found"})
+  }
+  await Blog.findByIdAndUpdate(req.params.id,req.body,{new : true})
+  .then(response => res.status(200).json(response))
+  .catch(error => next(error))
 })
 module.exports = blogRouter
